tests and refinement for specifying seed repositories for caches.
Restrict the possible seed caches to Irmin.basic rather than
Irmin.remote; we don't want to attempt to seed from a remote URI anyway
since it's potentially a chicken-and-egg problem,
and it will consistently fail in the case of in-memory stores
regardless.

Test to ensure we correctly report an error when attempting to pull from
a node in the repository which isn't parseable.

Also, some cleanup in test_arp.ml.